backend/requirements.txt
fastapi
uvicorn
pydantic
motor
pymongo
python-dotenv